+++ lib/chewy/query/pagination.rb
@@ -1,6 +1,10 @@
 module Chewy
   class Query
     module Pagination
+      # Returns request total found documents count
+      #
+      #   PlacesIndex.query(...).filter(...).total_count
+      #
       def total_count
         _response['hits'].try(:[], 'total') || 0
       end
@@ -8,9 +12,4 @@ module Chewy
   end
 end
 
-if defined?(::Kaminari)
-  require 'chewy/query/pagination/kaminari'
-  require 'chewy/query/pagination/kaminari_proxy'
-else
-  require 'chewy/query/pagination/proxy'
-end
+require 'chewy/query/pagination/kaminari' if defined?(::Kaminari)